+++ Source/CesiumRuntime/Public/Cesium3DTileset.h
@@ -14,7 +14,9 @@
 #include "CustomDepthParameters.h"
 #include "GameFramework/Actor.h"
 #include "Interfaces/IHttpRequest.h"
+#include "PrimitiveSceneProxy.h"
 #include <PhysicsEngine/BodyInstance.h>
+#include <atomic>
 #include <chrono>
 #include <glm/mat4x4.hpp>
 #include <unordered_map>
@@ -23,11 +25,14 @@
 
 class UMaterialInterface;
 class ACesiumCartographicSelection;
+class UCesiumBoundingVolumePoolComponent;
+class CesiumViewExtension;
 struct FCesiumCamera;
 
 namespace Cesium3DTilesSelection {
 class Tileset;
 class TilesetView;
+class TileOcclusionRendererProxyPool;
 } // namespace Cesium3DTilesSelection
 
 /**
@@ -37,6 +42,12 @@ class TilesetView;
 DECLARE_MULTICAST_DELEGATE_OneParam(
     FCesium3DTilesetLoadFailure,
     const FCesium3DTilesetLoadFailureDetails&);
+
+/**
+ * The delegate for the Acesium3DTileset::OnTilesetLoaded,
+ * which is triggered from UpdateLoadStatus
+ */
+DECLARE_DYNAMIC_MULTICAST_DELEGATE(FCompletedLoadTrigger);
 
 CESIUMRUNTIME_API extern FCesium3DTilesetLoadFailure
     OnCesium3DTilesetLoadFailure;
@@ -54,6 +65,9 @@ enum class ETilesetSource : uint8 {
    */
   FromUrl UMETA(DisplayName = "From Url")
 };
+
+UENUM(BlueprintType)
+enum class EApplyDpiScaling : uint8 { Yes, No, UseProjectDefault };
 
 UCLASS()
 class CESIUMRUNTIME_API ACesium3DTileset : public AActor {
@@ -156,6 +170,24 @@ private:
       Meta = (AllowPrivateAccess))
   ACesiumCreditSystem* ResolvedCreditSystem = nullptr;
 
+  /**
+   * The bounding volume pool component that manages occlusion bounding volume
+   * proxies.
+   */
+  UPROPERTY(
+      Transient,
+      BlueprintReadOnly,
+      Category = "Cesium",
+      Meta = (AllowPrivateAccess))
+  UCesiumBoundingVolumePoolComponent* BoundingVolumePoolComponent = nullptr;
+
+  /**
+   * The custom view extension this tileset uses to pull renderer view
+   * information.
+   */
+  TSharedPtr<CesiumViewExtension, ESPMode::ThreadSafe> _cesiumViewExtension =
+      nullptr;
+
 public:
   /** @copydoc ACesium3DTileset::CreditSystem */
   UFUNCTION(BlueprintCallable, Category = "Cesium")
@@ -209,6 +241,16 @@ public:
       Category = "Cesium|Level of Detail",
       meta = (ClampMin = 0.0))
   float MaximumScreenSpaceError = 16.0;
+
+  /**
+   * Scale Level-of-Detail by Display DPI. This increases the performance for
+   * mobile devices and high DPI screens.
+   */
+  UPROPERTY(
+      EditAnywhere,
+      BlueprintReadWrite,
+      Category = "Cesium|Level of Detail")
+  EApplyDpiScaling ApplyDpiScaling = EApplyDpiScaling::UseProjectDefault;
 
   /**
    * Whether to preload ancestor tiles.
@@ -389,6 +431,76 @@ public:
       meta = (EditCondition = "EnforceCulledScreenSpaceError", ClampMin = 0.0))
   float CulledScreenSpaceError = 64.0;
 
+  // This mirrors
+  // UCesiumRuntimeSettings::EnableExperimentalOcclusionCullingFeature so that
+  // it can be used as an EditCondition.
+  UPROPERTY(Transient, VisibleDefaultsOnly, Category = "Cesium|Tile Occlusion")
+  bool CanEnableOcclusionCulling = false;
+
+  /**
+   * Whether to cull tiles that are occluded.
+   *
+   * If this option is disabled, check that "Enable Experimental Occlusion
+   * Culling Feature" is enabled in the Plugins -> Cesium section of the Project
+   * Settings.
+   *
+   * When enabled, this feature will use Unreal's occlusion system to determine
+   * if tiles are actually visible on the screen. For tiles found to be
+   * occluded, the tile will not refine to show descendants, but it will still
+   * be rendered to avoid holes. This results in less tile loads and less GPU
+   * resource usage for dense, high-occlusion scenes like ground-level views in
+   * cities.
+   *
+   * This will not work for tilesets with poorly fit bounding volumes and cause
+   * more draw calls with very few extra culled tiles. When there is minimal
+   * occlusion in a scene, such as with terrain tilesets and applications
+   * focused on top-down views, this feature will yield minimal benefit and
+   * potentially cause needless overhead.
+   */
+  UPROPERTY(
+      EditAnywhere,
+      BlueprintGetter = GetEnableOcclusionCulling,
+      BlueprintSetter = SetEnableOcclusionCulling,
+      Category = "Cesium|Tile Occlusion",
+      meta = (EditCondition = "CanEnableOcclusionCulling"))
+  bool EnableOcclusionCulling = true;
+
+  /**
+   * The number of CesiumBoundingVolumeComponents to use for querying the
+   * occlusion state of traversed tiles.
+   *
+   * Only applicable when EnableOcclusionCulling is enabled.
+   */
+  UPROPERTY(
+      EditAnywhere,
+      BlueprintGetter = GetOcclusionPoolSize,
+      BlueprintSetter = SetOcclusionPoolSize,
+      Category = "Cesium|Tile Occlusion",
+      meta =
+          (EditCondition =
+               "EnableOcclusionCulling && CanEnableOcclusionCulling",
+           ClampMin = "0",
+           ClampMax = "1000"))
+  int32 OcclusionPoolSize = 500;
+
+  /**
+   * Whether to wait for valid occlusion results before refining tiles.
+   *
+   * Only applicable when EnableOcclusionCulling is enabled. When this option
+   * is enabled, there may be small delays before tiles are refined, but there
+   * may be an overall performance advantage by avoiding loads of descendants
+   * that will be found to be occluded.
+   */
+  UPROPERTY(
+      EditAnywhere,
+      BlueprintGetter = GetDelayRefinementForOcclusion,
+      BlueprintSetter = SetDelayRefinementForOcclusion,
+      Category = "Cesium|Tile Occlusion",
+      meta =
+          (EditCondition =
+               "EnableOcclusionCulling && CanEnableOcclusionCulling"))
+  bool DelayRefinementForOcclusion = true;
+
   /**
    * Refreshes this tileset, ensuring that all materials and other settings are
    * applied. It is not usually necessary to invoke this, but when
@@ -428,7 +540,16 @@ public:
       meta = (ShowOnlyInnerProperties, SkipUCSModifiedProperties))
   FBodyInstance BodyInstance;
 
+  /**
+   * A delegate that will be called whenever the tileset is fully loaded.
+   */
+  UPROPERTY(BlueprintAssignable, Category = "Cesium");
+  FCompletedLoadTrigger OnTilesetLoaded;
+
 private:
+  UPROPERTY(BlueprintGetter = GetLoadProgress, Category = "Cesium")
+  float LoadProgress = 0.0f;
+
   /**
    * The type of source from which to load this tileset.
    */
@@ -451,7 +572,7 @@ private:
       BlueprintSetter = SetUrl,
       Category = "Cesium",
       meta = (EditCondition = "TilesetSource==ETilesetSource::FromUrl"))
-  FString Url;
+  FString Url = "";
 
   /**
    * The ID of the Cesium ion asset to use.
@@ -584,11 +705,12 @@ private:
   bool EnableWaterMask = false;
 
   /**
-   * A custom Material to use to render this tileset, in order to implement
-   * custom visual effects.
+   * A custom Material to use to render opaque elements in this tileset, in
+   * order to implement custom visual effects.
    *
-   * The custom material should generally be created by copying the
-   * "M_CesiumOverlay" material and customizing it as desired.
+   * The custom material should generally be created by copying the Material
+   * Instance "MI_CesiumThreeOverlaysAndClipping" and customizing the copy as
+   * desired.
    */
   UPROPERTY(
       EditAnywhere,
@@ -598,15 +720,30 @@ private:
   UMaterialInterface* Material = nullptr;
 
   /**
+   * A custom Material to use to render translucent elements of the tileset, in
+   * order to implement custom visual effects.
+   *
+   * The custom material should generally be created by copying the Material
+   * Instance "MI_CesiumThreeOverlaysAndClippingTranslucent" and customizing the
+   * copy as desired. Make sure that its Material Property Overrides -> Blend
+   * Mode is set to "Translucent".
+   */
+  UPROPERTY(
+      EditAnywhere,
+      BlueprintGetter = GetTranslucentMaterial,
+      BlueprintSetter = SetTranslucentMaterial,
+      Category = "Cesium|Rendering")
+  UMaterialInterface* TranslucentMaterial = nullptr;
+
+  /**
    * A custom Material to use to render this tileset in areas where the
    * watermask is, in order to implement custom visual effects.
    * Currently only applicable for quantized-mesh tilesets that support the
    * water mask extension.
    *
-   * The custom material should generally be created by copying the
-   * "M_CesiumOverlayWater" material and customizing it as desired. For best
-   * results, any changes to the above material should also be duplicated in the
-   * Water Material.
+   * The custom material should generally be created by copying the Material
+   * Instance "MI_CesiumThreeOverlaysAndClippingAndWater" and customizing the
+   * copy as desired.
    */
   UPROPERTY(
       EditAnywhere,
@@ -628,6 +765,9 @@ protected:
   FString PlatformName;
 
 public:
+  UFUNCTION(BlueprintGetter, Category = "Cesium")
+  float GetLoadProgress() const { return LoadProgress; }
+
   UFUNCTION(BlueprintGetter, Category = "Cesium")
   ETilesetSource GetTilesetSource() const { return TilesetSource; }
 
@@ -658,6 +798,26 @@ public:
   UFUNCTION(BlueprintSetter, Category = "Cesium")
   void SetIonAssetEndpointUrl(const FString& InIonAssetEndpointUrl);
 
+  UFUNCTION(BlueprintGetter, Category = "Cesium|Tile Culling|Experimental")
+  bool GetEnableOcclusionCulling() const;
+
+  UFUNCTION(BlueprintSetter, Category = "Cesium|Tile Culling|Experimental")
+  void SetEnableOcclusionCulling(bool bEnableOcclusionCulling);
+
+  UFUNCTION(BlueprintGetter, Category = "Cesium|Tile Culling|Experimental")
+  int32 GetOcclusionPoolSize() const { return OcclusionPoolSize; }
+
+  UFUNCTION(BlueprintSetter, Category = "Cesium|Tile Culling|Experimental")
+  void SetOcclusionPoolSize(int32 newOcclusionPoolSize);
+
+  UFUNCTION(BlueprintGetter, Category = "Cesium|Tile Culling|Experimental")
+  bool GetDelayRefinementForOcclusion() const {
+    return DelayRefinementForOcclusion;
+  }
+
+  UFUNCTION(BlueprintSetter, Category = "Cesium|Tile Culling|Experimental")
+  void SetDelayRefinementForOcclusion(bool bDelayRefinementForOcclusion);
+
   UFUNCTION(BlueprintGetter, Category = "Cesium|Physics")
   bool GetCreatePhysicsMeshes() const { return CreatePhysicsMeshes; }
 
@@ -687,6 +847,14 @@ public:
 
   UFUNCTION(BlueprintSetter, Category = "Cesium|Rendering")
   void SetMaterial(UMaterialInterface* InMaterial);
+
+  UFUNCTION(BlueprintGetter, Category = "Cesium|Rendering")
+  UMaterialInterface* GetTranslucentMaterial() const {
+    return TranslucentMaterial;
+  }
+
+  UFUNCTION(BlueprintSetter, Category = "Cesium|Rendering")
+  void SetTranslucentMaterial(UMaterialInterface* InMaterial);
 
   UFUNCTION(BlueprintGetter, Category = "Cesium|Rendering")
   UMaterialInterface* GetWaterMaterial() const { return WaterMaterial; }
@@ -735,6 +903,8 @@ public:
   virtual void EndPlay(const EEndPlayReason::Type EndPlayReason) override;
   virtual void PostLoad() override;
   virtual void Serialize(FArchive& Ar) override;
+
+  void UpdateLoadStatus();
 
   // UObject overrides
 #if WITH_EDITOR
@@ -834,6 +1004,10 @@ private:
    * was given in the root tile.
    */
   void OnFocusEditorViewportOnThis();
+
+  void RuntimeSettingsChanged(
+      UObject* pObject,
+      struct FPropertyChangedEvent& changed);
 #endif
 
 private:
@@ -846,10 +1020,13 @@ private:
   uint32_t _lastTilesLoadingLowPriority;
   uint32_t _lastTilesLoadingMediumPriority;
   uint32_t _lastTilesLoadingHighPriority;
+  bool _activeLoading;
 
   uint32_t _lastTilesVisited;
   uint32_t _lastCulledTilesVisited;
   uint32_t _lastTilesCulled;
+  uint32_t _lastTilesOccluded;
+  uint32_t _lastTilesWaitingForOcclusionResults;
   uint32_t _lastMaxDepthVisited;
 
   std::chrono::high_resolution_clock::time_point _startTime;
@@ -859,6 +1036,8 @@ private:
   bool _beforeMoviePreloadSiblings;
   int32_t _beforeMovieLoadingDescendantLimit;
   bool _beforeMovieKeepWorldOriginNearCamera;
+
+  bool _scaleUsingDPI;
 
   // This is used as a workaround for cesium-native#186
   //