+++ Source/CesiumRuntime/Public/CesiumCreditSystem.h
@@ -2,11 +2,13 @@
 
 #pragma once
 
+#include "Components/WidgetComponent.h"
 #include "Engine/Blueprint.h"
 #include "GameFramework/Actor.h"
 #include "UObject/Class.h"
 #include "UObject/ConstructorHelpers.h"
 #include <memory>
+#include <unordered_map>
 
 #include "CesiumCreditSystem.generated.h"
 
@@ -29,26 +31,16 @@ public:
 
   ACesiumCreditSystem();
 
-  /**
-   * The credits text to display.
-   */
-  UPROPERTY(BlueprintReadOnly, Category = "Cesium")
-  FString Credits = "";
+  void BeginPlay() override;
 
-  UPROPERTY(BlueprintReadOnly, Category = "Cesium")
-  FString OnScreenCredits = "";
+  UPROPERTY(EditDefaultsOnly, Category = "Cesium")
+  TSubclassOf<UUserWidget> CreditsWidgetClass;
 
   /**
    * Whether the credit string has changed since last frame.
    */
   UPROPERTY(BlueprintReadOnly, Category = "Cesium")
   bool CreditsUpdated = false;
-
-  /**
-   * Whether any credits are part of the frame.
-   */
-  UPROPERTY(BlueprintReadOnly, Category = "Cesium")
-  bool DisplayCredits = false;
 
   // Called every frame
   virtual bool ShouldTickIfViewportsOnly() const override;
@@ -62,6 +54,9 @@ public:
 private:
   static UClass* CesiumCreditSystemBP;
 
+  UPROPERTY()
+  class UScreenCreditsWidget* _creditsWidget;
+
   /**
    * A tag that is assigned to Credit Systems when they are created
    * as the "default" Credit System for a certain world.
@@ -72,4 +67,7 @@ private:
   std::shared_ptr<Cesium3DTilesSelection::CreditSystem> _pCreditSystem;
 
   size_t _lastCreditsCount;
+
+  FString ConvertHtmlToRtf(std::string html);
+  std::unordered_map<std::string, FString> _htmlToRtf;
 };